src/requirements.txt
asn1crypto==1.0.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.4
django-configurations==2.1
django-js-asset==1.2.2
django-mptt==0.10.0
djangorestframework==3.10.3
future==0.17.1
idna==2.8
importlib-metadata==0.23
kombu==4.6.5
ldap3==2.6
more-itertools==7.2.0
psycopg2==2.7.5
psycopg2-binary==2.8.3
django-decouple
django-cors-headers==3.2.0
drf-yasg